+++ Addons/Fbx.cs
@@ -1,9 +1,11 @@
-// todo Once we drop support for 2018.3, use optional assembly definitions
-using System;
+#if FBX_EXPORTER
+
 using UnityEditor;
 using System.Reflection;
 using System.Linq;
+using UnityEditor.Formats.Fbx.Exporter;
 using UnityEditor.ProBuilder;
+using UnityEditor.ProBuilder.Actions;
 
 namespace UnityEngine.ProBuilder.Addons.FBX
 {
@@ -26,29 +28,7 @@ namespace UnityEngine.ProBuilder.Addons.FBX
     [InitializeOnLoad]
     static class Fbx
     {
-        private static Assembly FbxExporterAssembly
-        {
-            get
-            {
-                try
-                {
-                    return Assembly.Load("Unity.Formats.Fbx.Editor");
-                }
-                catch (System.IO.FileNotFoundException)
-                {
-                    return null;
-                }
-            }
-        }
-        
-        static readonly Type[] k_ProBuilderTypes = new Type[]
-        {
-            typeof(BezierShape),
-            typeof(PolyShape),
-            typeof(Entity)
-        };
-
-        static FbxOptions m_FbxOptions = new FbxOptions() {
+        static FbxOptions s_FbxOptions = new FbxOptions() {
             quads = true
         };
 
@@ -59,52 +39,35 @@ namespace UnityEngine.ProBuilder.Addons.FBX
 
         static void TryLoadFbxSupport()
         {
-            if (FbxExporterAssembly == null)
-            {
-                return;
-            }
-
-            var modelExporter = FbxExporterAssembly.GetType("UnityEditor.Formats.Fbx.Exporter.ModelExporter");
-            var registerMeshCallback = modelExporter.GetMethods(BindingFlags.NonPublic | BindingFlags.Static).Where(x => x.Name == "RegisterMeshCallback").First(x => x.ContainsGenericParameters);
-            registerMeshCallback = registerMeshCallback.MakeGenericMethod(typeof(ProBuilderMesh));
-
-            var getMeshForComponent = FbxExporterAssembly.GetTypes()
-               .Where(t => t.BaseType == typeof(MulticastDelegate) && t.Name.StartsWith("GetMeshForComponent"))
-               .First(t => t.ContainsGenericParameters);
-            
-            getMeshForComponent = getMeshForComponent.MakeGenericType(typeof(ProBuilderMesh));
-            var meshDelegate = Delegate.CreateDelegate(getMeshForComponent, typeof(Fbx).GetMethod("GetMeshForComponent", BindingFlags.NonPublic | BindingFlags.Static));
-
-            registerMeshCallback.Invoke(null, new object[] { meshDelegate, true });
-            
-            m_FbxOptions.quads = ProBuilderSettings.Get<bool>("Export::m_FbxQuads", SettingsScope.User, true);
+            ModelExporter.RegisterMeshCallback<ProBuilderMesh>(GetMeshForPBComponent, true);
+            s_FbxOptions.quads = ProBuilderSettings.Get<bool>("Export::m_FbxQuads", SettingsScope.User, true);
         }
 
-        static bool GetMeshForComponent(object exporter, ProBuilderMesh pmesh, object node)
+        static bool GetMeshForPBComponent(ModelExporter exporter, ProBuilderMesh pmesh, Autodesk.Fbx.FbxNode node)
         {
             Mesh mesh = new Mesh();
-            MeshUtility.Compile(pmesh, mesh, m_FbxOptions.quads ? MeshTopology.Quads : MeshTopology.Triangles);
+            MeshUtility.Compile(pmesh, mesh, s_FbxOptions.quads ? MeshTopology.Quads : MeshTopology.Triangles);
 
-            // using reflection to call: exporter.ExportMesh(mesh, node, pmesh.GetComponent<MeshRenderer>().sharedMaterials)
             var pMeshRenderer = pmesh.GetComponent<MeshRenderer>();
             var sharedMaterials = pMeshRenderer ? pMeshRenderer.sharedMaterials : null;
-            var exportMeshMethod = exporter.GetType().GetMethod("ExportMesh", BindingFlags.NonPublic | BindingFlags.Instance, null, new Type[] { typeof(Mesh), node.GetType(), typeof(Material[]) }, null);
-            exportMeshMethod.Invoke(exporter, new object[] { mesh, node, sharedMaterials });
+
+            exporter.ExportMesh(mesh, node, sharedMaterials);
 
             Object.DestroyImmediate(mesh);
 
-            // probuilder can't handle mesh assets that may be externally reloaded, just strip pb stuff for now.
-            foreach (var type in k_ProBuilderTypes)
-            {
-                var component = pmesh.GetComponent(type);
-                if (component != null)
-                    Object.DestroyImmediate(component);
-            }
+            //Need to have ExportOptions accessible to remove this reflection
+            var exporterType = exporter.GetType().GetMethods(BindingFlags.NonPublic | BindingFlags.Instance)
+                .First(x => x.Name == "get_ExportOptions").Invoke(exporter, null).GetType();
 
-            pmesh.preserveMeshAssetOnDestroy = true;
-            Object.DestroyImmediate(pmesh);
+            if(exporterType == typeof(ConvertToPrefabSettingsSerialize))
+            {
+                // probuilder can't handle mesh assets that may be externally reloaded, just strip pb stuff for now.
+                StripProBuilderScripts.DoStrip(pmesh);
+            }
 
             return true;
         }
     }
 }
+
+#endif