+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [5.0.4] - 2021-06-08
+
+### Bug Fixes
+
+- [case: 1334017] Fixed errors while exporting a PBShape using FBX Exporter and cleaning export.
+- [case: 1332226] Fixed issue where some Gizmos menu items would be missing in projects that have ProBuilder package installed.
+- [case: 1324374] Fixed incorrect vertex/edge/face rect selection when mesh's parent is rotated and/or scaled.
+
 ## [5.0.3] - 2021-04-01
 
 ### Bug Fixes